pyproject.toml
[project]
name = "ria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "numpy>=2.4.0",
    "openai-agents>=0.6.4",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "sounddevice>=0.5.3",
]
